Spliting the logic into separate files
